Refactor Footer: In prep for fixed controls
